[Android] Fix usage of Import Password localisation string substitution
When we fetch localisation strings which have different cases for
singular and plural, we should pass the count twice.

This is because the first argument determines whether the singular or
plural form will be used and the second one is used for substitution.

Technically, we could hard-code the first count param as a numeric
literal, but this gains us nothing and it is the opinion of the author
that using the variable as a single-source-of-truth is the better
approach.

Resolves brave/brave-browser#50245