[Android] Fix VPN city selection UI to reflect an option choice (#31976)
Previously in af14cfe912afadcccef1d958a602a657782ff33a we fixed the
server selection dialog and prevented dialog closure when the user
selects an option.

As a result, the city selection dialog also needed code added to ensure
that it updated when the user selected an option. This commit implements
that change in the same manner as the aforementioned one.

Resolves brave/brave-browser#50159