[wallet][safe-numerics] Improve `BtcLikeSerializerStream` use (#31886)
This PR greatly simplifies `BtcLikeSerializerStream`, but making the
vector used as a buffer part of the class itself. This lines up with how
all uses of this class occur, with the vector created just before the
serialiser, so making the vector a member of the serialiser simplifies
things significantly.

A few `checked_cast`s have been added to avoid future mistakes around
this particular area of the code.

This change also rolls out the use of base::StrictNumeric in certain
places where this is approrpriate.

Bug https://github.com/brave/brave-browser/issues/50271