[plaster] Add `chrome/app/BUILD.gn` 🩹 (#29734)
This change adds a  🩹 for the patching we do in `chrome/app/BUILD.gn`.
The output is the same as the patch we already have in place, so nothing
changes there. This particular plaster is meant to be flexible, so to
not break by the lines around changing. It is designed to add the line
just at the end of the block, regardless of what may have changed around
it.

```
grit_strings("generated_resources") {
  ...
  ...
  ...
  deps = [ "//brave/app:brave_generated_resources_grit" ]
}
```

Resolves https://github.com/brave/brave-browser/issues/47088