[Android] Apply tinting for Import/Export password manager icons
Previously, the import/export icons were not correctly tinted when the
Android OS was set to dark mode.

To correct this, the icons no longer use a hard-coded fill colour and an
iconTint property is applied in order to leverage the built-in support
that Chromium has for automatically adjusting icons for light/dark mode.

The icons have also been replaced with the upload/download SVG paths
from the brave/leo repo.

Resolves brave/brave-browser#50232