[CodeHealth] Convert several `reinterpret_cast` to `span` uses
This PR converts several places where `reinterpret_cast` was being used,
to now use `base::span`. In several cases, the use of
`base::as_string_view` is enough to avoid the cast. In some other
places, the use of `base::span` make the buffer ranged, and easier to
think about.

Bug: https://github.com/brave/brave-browser/issues/41661