[unsafe-buffers] Spanify `BraveSessionCache::GenerateRandomString` (#32137)
This function has still been lagging behind significantly. This change
removes one of the UNSAFE_TODOs in this function. It also removes a
`reinterpret_cast<uint64_t>` that has a potential for misaligned reads
(although it is not very clear if it could happen, but this was reported
in BraveSessionCache::GenerateRandomString).

Reported in https://hackerone.com/reports/3406392.

Bug: https://github.com/brave/brave-browser/issues/41661