Updated margin between vertical tab and contents (#32118)
Resolves brave/brave-browser#50557
Resolves brave/brave-browser#50558

We'll have 1px right(or left) inset for vertical tab when rounded corners is off or floating mode
for drawing 1px border line.

We'll have -2px(tabs::kMarginForVerticalTabContainers/2) inset for vertical tab when rounded corners is on.
With rounded corners, we need fixed distance(tabs::kMarginForVerticalTabContainers)
between vertical tab and contents. Each side has half of that as a margin to draw contents shadow.

The distance between them will not be changed even we have different contents margin thickness.
We'll update contents margin(#31583) and it should not affect
the distance between them.

TEST=VerticalTabStripBrowserTest.VisualState, BraveBrowserViewTest.LayoutWithVerticalTabTest