[ShieldsUI] Reland: Add tab switcher for page info bubble (#32143)
Reland of #32026.

In browser tests, scoped observation of BraveShieldsTabHelper was
being reset after the web contents (and the associated
BraveShieldsTabHelper) was destroyed.

In this change, AddObserver and RemoveObserver are now called
directly, since there is not currently any event which indicates
when a BraveShieldsTabHelper instance is being destroyed, and it
is not safe to assume when its lifetime ends otherwise.