Remove BraveDownloadCommands
We don't want to add extra deps to //brave/browser from //chrome/browser.

Also the methods in BraveDownloadCommands, such as IsCommandEnabled and
ExecuteCommand, should not be handled in a DownloadCommands. As they
are handled in DownloadUIModel and DownloadItemModel(which is a
implementation of it on Desktop).

So moving the handling logic to the DownloadItemModel by replacing
upstream implementation with Brave's one.