Fixed network audit failure from feature enabled state change
The number of built-in panel could be different as some panel
could be hidden if its feature is disabled.
Due to this, network audit check failed as this test only
checkes total count with constant(4). Fixed by deleting this number checking.
It's added to increase total test time when we have more panels.
We'll f/u test total time calculation.