[Wallet] [Polkadot] add chain_getHeader rpc call (#32159)
Resolves brave/brave-browser#50359

Add an RPC call for chain_getHeader which retrieves the block for the specified block hash or, if not provided, the latest block in the chain. Right now we only extract the block number and the parent hash, leaving the extrinsics root and state root untouched along with the digest/logs.