Allow vertical tab strip to be opened when window is inactive
As per user request, removes the restriction that prevents the vertical
tab strip from expanding when the browser window is inactive.
Previously, we didn't schedule the expanding of the tab strip if the
window wasn't active. But this required users to activate the window
first before they could find tabs they wanted to see. This restriction
is not necessary, so we remove it to improve user experience.