[Tabs] Add pref to always hide tab close button (#32189)
* Add pref to always hide tab close button

Some users prefer a cleaner tab appearance without close buttons, as
they close tabs using other methods (e.g., middle-click, shortcuts).
So close button could be no more than a visual distraction or accidental
click.

This commit introduces a new preference that allows users to always hide
the tab close button.

* Introduce ControllableCloseButtonState

This is a sort of decorator class that wraps bool showing_close_button_
so that we can consider additional factors when determining whether to show
the close button, such as whether the mouse is hovering over the tab,
or whether the controller wants to always hide the close button.