add chain_getBlockHash RPC call (#32316)
Resolves brave/brave-browser#50461

Add an RPC call for getting the hash of a block's header. This is basically required for getting the genesis hash and it'll also enable us to skip hashing the potentially-large headers by just dispatching to a network call which can be parallelized. We will need the blockhash of the block we're designating as the start of the mortality period.