[iOS] [CodeHealth] Remove `BraveSkusSDK` wrapper type (#32431)
The `BraveSkusSDK` class simply hosted a `SkusService` and added utility methods to handle `BraveStoreProduct` types in the skus service. This change removes it in favour of simply adding these methods directly on `SkusService`

This change also:
- Removes up an empty method: `refreshAllSkusOrders`
- Deprecates the `BraveStoreSDK` singleton
- Moves the `receipt(for:)` method into the `AppStoreReceipt` type

There is no logical change with this removal, all places that used `BraveSkusSDK.shared` now grab a `SkusService` directly from the factory the same way `BraveSkusSDK` did.