[cr143] Fixed browser UI is not shown in fullscreen on macOS (#32488)
Resolves brave/brave-browser#51008

This is an upstream fix. It will go away once upstream uplifts it to M143.

Chromium change:

https://chromium-review.googlesource.com/c/chromium/src/+/7107360

[Browser Frame] Abandon reading Mac caption button positions

These were not reliable during fullscreen transitions in a way that was
not possible to accurately filter out. This now uses the old fallback
hard-coded constants that were being used in all PWA windows anyway.