[Polkadot] [Wallet] add unsigned transfer_allow_death extrinsic (#32290)
Add our first required extrinsic which enables sending DOT.

The various Polkadot documents will mention that transfer_allow_death is the preferred pallet call for making DOT transfers and automatically reaps accounts whose DOT value falls below the existential deposit value.

We do not innovate in this space, instead essentially mimicking what the polkadot-js/api implementation does which is create a simple hex string with the type of the extrinsic call and the recipient address along with the amount of funds to send.

We have Encode() and Decode() which allow us to serialize and deserialize the unsigned transaction so we can safely store it as a hex string anywhere we'd like.

We also introduce a new Rust-based data structure to store chain runtime metadata. We do this because the binary blobs returned from the RPC nodes will be heavily SCALE-encoded, which means we need Rust regardless. Keeping everything in Rust simplifies development flow and we have the extensibility to add this metadata parsing code at a later point in the future when our hard-coded indices inevitably prove insufficient.