[Leo] Extract OAI completion response parsing into a share function (#32661)
This PR is in preparation to share this same util function in
conversation V2 API client. Before this PR, we only support parsing
`model` in the response for conversation API and not custom models.
This PR does not change this fact so we only support looking up Leo
models from `model` in the response which will be used by conversation
V2 and do not support looking up custom models, which we can add
support in the future if needed.