[ads] Send platform, masked version number, and build channel in confirmation payload
Because the user agent string is no longer sent for OHTTP, we need an
alternative way to get the browser version, platform and build channel.
This change adds the platform, masked version number, and build channel
for non-Rewards users to the payload so that confirmation logic can
continue to function correctly.