[Leo] Implement GenerateConversatinoTitle in conversation API v2 (#32773)
* Implement GenerateConversatinoTitle in conversation API v2

This mostly follows our OAI implementation for title generation, where
we would send a separate request to get title after first assistant
response is back. This request would have 1 message with page content
blocks, a page excerpt block (if there is selected text), and a content
block with kRequestTitle type and first user turn text or first
assistant response text as the text content.

There is a difference between existing OAI implementation and the one
in oai_message_utils which we will put selected text in a page excerpt
block instead of mixing it with user turn text in oai_message_utils.

There is a small refactoring which we pull out a new util function
BuildOAIPageContentBlocks similar to what we have in
EngineConsumerOAIRemote::BuildPageContentMessages so we can share the
logic for building page content blocks more inside oai_message_utils.