[iOS] Shields Migration from CoreData to Content Settings (#32569)
* Migrate Shields settings from CoreData `Domain` object to Content Settings.
* Add migration unit tests.
* Update migration to option 2. This will include `www|m|mobile` in migrations to
account for `domainURL` on iOS treating them as the same site (which differs
from content settings treating them as separate).
* Adjust logic so we continue to update content settings when feature flag is
disabled, but migrations have already occurred.
This could happen if we roll out the feature flag using percentages as there
is no guarantee that a user in the first study at say 25%, would be included
in the follow-up studies.