[iOS] check selected tab's mode matches current mode (#32850)
changed the check for selected tab shred availability

Problem: Selected tab still references the last selected tab after transitioning from one browsing mode to the other. So checking for `isShredAvailable` will always return true if the previous mode had a selected tab.

Solution: Ensure that the selected tab's mode matches the active mode of the browser in addition to checking for if shred is available.

Signed-off-by: Eli Hini <ehini@brave.com>