Browser Search Query Analytics
This feature measures search activity in the Brave browser in a
privacy-preserving way by sending a lightweight analytics ping when a
user views search results, without ever collecting the search query. The
ping contains a payload with key-value pairs and is sent over OHTTP,
ensuring network-level unlinkability. To reduce the risk of fingerprinting,
all values reported are picked from a limited set of options.

Co-authored-by: Aleksei Seren <aseren@brave.com>