Return nullopt when completion content is empty in ParseOAICompletionResponse
It shouldn't lead to a completion event with empty string in
conversation API cases, as we currently see responses with tool_calls
where content is empty string. This would just make v2 follow the v1
behavior. For OAI, we return empty completion event in OnQueryCompleted
anyway, so it doesn't change the behavior. The behavior change on OAI is
we wouldn't fire completion event in OnQueryDataReceived if content is
empty, but it does make sense in this case as there is no point fire it.

Also Extract GetOAIContentContainer for future usage.