[Polkadot] [Wallet] update PolkadotWalletService to fetch metadata (#32859)
Resolves brave/brave-browser#51520

We begin to outfit the PolkadotWalletService with sequence checking, as most of
Wallet seems to rely on running in an implicit sequence and the test environment
is able to recreate this fine.

We update the constructor to make two simultaneous network calls, to both the
mainnet and testnet, to build the required chain metadata that'll store the
metadata required to build extrinsics in a way that's accessible from the
front-end via mojom.