Add pref to control the visibility of the sharing hub icon address bar
This commit introduces a new preference to control the visibility of
the Sharing Hub icon in the address bar.

Changes made in this commit:
- Registered the new pref in `browser_ui_prefs.h/cc` and ensured its
  integration into `pref_names.h`.
- Updated Chromium and Brave-side files to conditionally show or hide
  the Sharing Hub icon based on the new pref value.
  - SharingHubIconView: Override the UpdateImpl method to observe the
    pref change.
  - SharingHubBubbleControllerDesktopImpl: Override the
    ShouldOfferOmniboxIcon method to consult the pref value.
- Modified `list_action_modifiers.cc` and related UI code (mojom, views)
  to show to pref value in "Customize Toolbar" side panel.

This change makes it possible for users (or internal policies) to hide or show the Sharing Hub address bar icon without affecting overall sharing functionality, supporting UI customization and policy compliance.