Enable Video PIP controls update 2024 by default(~cr143/1.86.x) (#32958)
Enable Video PIP controls update 2024 by default

The Video PIP controls update 2024 contains a security critical feature:
shows origin of the video in the PIP window's title.

As of cr144, this feature is adopted by default on our side, but the
previous version of Brave still uses the old version of the feature.

In order to align with the latest security critical feature, this commit
enables the feature by default (~cr143/1.86.x).